requirements.txt
# Core dependencies
streamlit>=1.28.0,<2.0.0
requests>=2.31.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0

# Text processing and matching
fuzzywuzzy>=0.18.0,<1.0.0
python-levenshtein>=0.21.0,<1.0.0
